Intelligently set initial probabilities of propositions
If a bunch of props are mutually exclusive set their initial probabilities to zero. That means one flip will satisfy the clause, whereas, if the initial probabilities are .5, you start with half of them true and you have to blindly do a lot of flips until you're down to just one.